% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aquamaps_species}
\alias{aquamaps_species}
\title{aquamaps_species}
\format{
An object of class \code{list} of length 11.
}
\source{
See aquamaps.R
}
\usage{
aquamaps_species
}
\description{
aquamaps_species
}
\keyword{datasets}
