% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csv_check.R
\name{check_translation_csv}
\alias{check_translation_csv}
\title{csv_check}
\usage{
check_translation_csv(path)
}
\value{
The return value, if any, from executing the utility.
}
\description{
Check the CSV written by hand
}
