% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{computeEffectiveForOneModel}
\alias{computeEffectiveForOneModel}
\title{Compute effective for 1 model}
\usage{
computeEffectiveForOneModel(
  model,
  currentYear,
  results,
  generationtime,
  nbCohorts,
  years,
  parameter,
  cohortWeight,
  spawnersTo_50
)
}
\arguments{
\item{model}{the climatic model name}

\item{currentYear}{current time step}

\item{results}{a list with model outputs}

\item{generationtime}{number of years contributing to a spawner run}

\item{years}{all years used in the simulation}

\item{cohortWeight}{weight of cohorts participating to reproduction}

\item{spawnersTo_50}{spawner number that results in half of the spawners participating in spawning}

\item{parameters}{all parameters used in simulation}
}
\description{
Compute effective for 1 model
}
