% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_b_first_fct_sql_q.R
\name{ecosystem_table}
\alias{ecosystem_table}
\title{Title}
\usage{
ecosystem_table(
  species,
  case_study,
  ecosystem,
  r,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{species, case_study, ecosystem}{Param for file in SQL}

\item{r}{reactiveValues}

\item{session}{Shiny session object}
}
\description{
Title
}
