% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_services_data.R
\docType{data}
\name{fake_services_data}
\alias{fake_services_data}
\title{Fake dataset for services}
\format{
A data frame
\describe{
  \item{watershed_code}{code of watershed}
  \item{species_code}{Code of species}
  \item{service_code}{Code of service}
  \item{value}{value of service}
}
}
\usage{
fake_services_data
}
\description{
Fake dataset for services
}
\keyword{datasets}
