% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_watershed_data.R
\docType{data}
\name{fake_watershed_data}
\alias{fake_watershed_data}
\title{Fake dataset for watershed names and positions}
\format{
A data frame
\describe{
  \item{name}{name of watershed}
  \item{lat}{Latitude in geographic coordinates}
  \item{long}{Longitude in geographic coordinates}
}
}
\usage{
fake_watershed_data
}
\description{
Fake dataset for watershed names and positions
}
\keyword{datasets}
