% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frontiers}
\alias{frontiers}
\title{frontiers}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 35 rows and 2 columns.
}
\source{
See data-raw
}
\usage{
frontiers
}
\description{
frontiers
}
\keyword{datasets}
