% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_species.R
\name{get_conservation_status}
\alias{get_conservation_status}
\title{Title}
\usage{
get_conservation_status(species_id, con)
}
\arguments{
\item{species_id}{}

\item{con}{}
}
\value{

}
\description{
Title
}
