% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_simulation.R
\name{get_data_simulation}
\alias{get_data_simulation}
\title{Get data ready for simulations on page 4}
\usage{
get_data_simulation(conn_eurodiad)
}
\arguments{
\item{conn_eurodiad}{Connection to database}
}
\value{
list of tables necessary for simulations
}
\description{
Get data ready for simulations on page 4
}
\examples{
\dontrun{
data_simulation <- get_data_simulation(conn_eurodiad)
data_simulation[["data_catchment"]]
data_simulation[["outlet_distance"]]
data_simulation[["hydiad_parameter"]]
data_simulation[["data_hsi_nmax"]]
data_simulation[["reference_results"]]
data_simulation[["data_ni0"]]
}
}
