% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_help_bubble_entries}
\alias{get_help_bubble_entries}
\title{Returns all help bubble entries}
\usage{
get_help_bubble_entries()
}
\value{
A character vector.
}
\description{
Returns all help bubble entries
}
