% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_second_fct_query_and_plot.R
\name{get_hybrid_model}
\alias{get_hybrid_model}
\title{Title}
\usage{
get_hybrid_model(
  species_id,
  scenario,
  session = shiny::getDefaultReactiveDomain()
)
}
\description{
Description
}
