% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{get_model_nit}
\alias{get_model_nit}
\title{Get Nit results from model results}
\usage{
get_model_nit(results)
}
\arguments{
\item{results}{List of results}
}
\value{
A list of Nit per model
}
\description{
Get Nit results from model results
}
