% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mongo.R
\name{launch_mongo}
\alias{launch_mongo}
\title{mongo}
\usage{
launch_mongo(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session}
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
