% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{nit_feature}
\alias{nit_feature}
\title{Calculate NIT feature}
\usage{
nit_feature(data_list)
}
\arguments{
\item{data_list}{List of Nit outputs only}
}
\value{
A tibble
}
\description{
Calculate NIT feature
}
