% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{nit_feature_species}
\alias{nit_feature_species}
\title{Get nit reference and predictions for one species and all basins}
\usage{
nit_feature_species(Nit_list, reference_results, selected_latin_name)
}
\arguments{
\item{Nit_list}{Nit_list as issued from [get_model_nit()]}

\item{reference_results}{reference_results as issued from [prepare_datasets()]}

\item{selected_latin_name}{Latin species name}
}
\value{
data.frame of Nit results for one species and one basin
}
\description{
Get nit reference and predictions for one species and all basins
}
