% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{nit_feature_species_basin}
\alias{nit_feature_species_basin}
\title{Get nit reference and predictions for one species and one basin}
\usage{
nit_feature_species_basin(
  Nit_list,
  reference_results,
  selected_latin_name,
  basin
)
}
\arguments{
\item{Nit_list}{Nit_list as issued from [get_model_nit()]}

\item{reference_results}{reference_results as issued from [prepare_datasets()]}

\item{selected_latin_name}{Latin species name}

\item{basin}{Name of the basin}
}
\value{
data.frame of Nit results for one species and one basin
}
\description{
Get nit reference and predictions for one species and one basin
}
