% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{runSimulation}
\alias{runSimulation}
\title{Run simulation}
\usage{
runSimulation(
  selected_latin_name,
  hydiad_parameter,
  anthropogenic_mortality,
  catchment_surface,
  data_hsi_nmax,
  data_ni0,
  outlet_distance,
  scenario = "rcp85",
  verbose = FALSE
)
}
\arguments{
\item{selected_latin_name}{Species latin name}

\item{hydiad_parameter}{Hydiad model parameters}

\item{anthropogenic_mortality}{table of anthropogenic mortalities}

\item{catchment_surface}{Surface of basins}

\item{data_hsi_nmax}{HSI Nmax values}

\item{data_ni0}{ni0 values}

\item{outlet_distance}{distance from outlet}

\item{scenario}{Climatic scenario. e.g. "rcp85"}

\item{verbose}{Logical.}
}
\value{
List of models outputs
}
\description{
Run simulation
}
