% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_a_fourth_fct_map.R
\name{tm_draw}
\alias{tm_draw}
\title{Title}
\usage{
tm_draw(
  species_latin_name,
  spatial_type,
  con,
  yearStart = 2003,
  yearEnd = 2017,
  dataCatchment,
  catchment_geom,
  dataALL,
  ices_geom
)
}
\arguments{
\item{species_latin_name}{}

\item{spatial_type}{}

\item{con}{}

\item{yearStart}{}

\item{yearEnd}{}

\item{dataCatchment}{}

\item{catchment_geom}{}

\item{dataALL}{}

\item{ices_geom}{}
}
\value{

}
\description{
Title
}
