% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tm_frontiers}
\alias{tm_frontiers}
\title{tm_frontiers}
\format{
An object of class \code{tmap} of length 3.
}
\source{
See data-raw
}
\usage{
tm_frontiers
}
\description{
tm_frontiers
}
\keyword{datasets}
