% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_data.R
\docType{data}
\name{World}
\alias{World}
\title{World map}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 177 rows and 16 columns.
}
\source{
\url{https://www.naturalearthdata.com/} for \code{World}

\url{https://happyplanetindex.org/} for \code{World}
}
\usage{
data(World)
}
\description{
Maps of the world , class \code{\link[sf:sf]{sf}}
}
\details{
\code{World} World map. The default projection for this world map is Eckhart IV since area sizes are preserved, which is a very important property for choropleths.
}
\keyword{datasets}
