% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_csv_check.R
\name{check_translation_csv}
\alias{check_translation_csv}
\title{csv_check}
\usage{
check_translation_csv(path, source_encoding = "latin1", new_path = path)
}
\arguments{
\item{path}{path to file to check}

\item{source_encoding}{Encoding of the source file}

\item{new_path}{Where to save the modified file}
}
\value{
The return value, if any, from executing the utility.
}
\description{
Check the CSV written by hand
}
