% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_b_first_fct_sql_q.R
\name{ecosystem_table}
\alias{ecosystem_table}
\title{Title}
\usage{
ecosystem_table(
  species,
  case_study,
  ecosystem,
  r,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{species_latin_name}{The latin name of the species}

\item{spatial_type}{Geom to use in the map}

\item{con}{The Connection object}

\item{yearStart, yearEnd}{date used}

\item{dataCatchment, catchment_geom, dataALL, ices_geom}{internal datasets}
}
\description{
Title
}
