% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_c_third_fct_query_and_plot.R
\name{plot_hsi_nit}
\alias{plot_hsi_nit}
\alias{plot_nit}
\alias{plot_hsi}
\title{Plot hsi and nit on the same graph}
\usage{
plot_hsi_nit(
  model_res,
  selected_year,
  selected_bv,
  lg,
  withNitStandardisation = FALSE
)

plot_nit(
  model_res_filtered,
  selected_year,
  lg,
  withNitStandardisation = FALSE,
  with_colour_source = NULL
)

plot_hsi(model_res_filtered, selected_year, lg)
}
\arguments{
\item{model_res}{model outputs}

\item{selected_year}{Numeric. selected_year}

\item{selected_bv}{Character. Selected BV}

\item{lg}{lang}

\item{withNitStandardisation}{Logical. Whether to standardise NIT.}

\item{with_colour_source}{Name of a variable to change colour lines}
}
\description{
Plot hsi and nit on the same graph

Plot NIT only

Plot NIT only
}
