% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_d_fourth_fct.R
\name{runSimulation}
\alias{runSimulation}
\title{Run simulation}
\usage{
runSimulation(
  selected_latin_name,
  hydiad_parameter,
  anthropogenic_mortality,
  catchment_surface,
  data_hsi_nmax,
  data_ni0,
  outlet_distance,
  verbose = FALSE
)
}
\arguments{
\item{selected_latin_name}{}

\item{hydiad_parameter}{}

\item{anthropogenic_mortality}{}

\item{catchment_surface}{}

\item{data_hsi_nmax}{}

\item{data_ni0}{}

\item{outlet_distance}{}

\item{verbose}{}
}
\description{
Run simulation
}
