% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_a_first_fct_map.R
\name{tm_draw}
\alias{tm_draw}
\title{Title}
\usage{
tm_draw(
  species_latin_name,
  spatial_type,
  con,
  yearStart = 2003,
  yearEnd = 2017,
  dataCatchment,
  catchment_geom,
  dataALL,
  ices_geom,
  ices_division,
  positive_catch_area,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{species_latin_name}{The latin name of the species}

\item{spatial_type}{Geom to use in the map}

\item{con}{The Connection object}

\item{yearStart, yearEnd}{date used}

\item{dataCatchment, catchment_geom}{internal datasets for continental waters}

\item{dataALL, ices_geom, ices_division, positive_catch_area}{internal datasets for marines water}

\item{session}{The Shiny Session object}
}
\value{
A tmap object
}
\description{
Title
}
